+++ crates/crcbl-render/src/contact_shadows.rs
@@ -246,7 +246,7 @@ impl ContactShadows {
             .execute(move |ctx| {
                 let view = ctx.image_view(depth);
                 let device = ctx.device();
-                let entries = vec![
+                let entries = [
                     BindGroupEntry {
                         binding: 0,
                         array_index: 0,
@@ -267,7 +267,7 @@ impl ContactShadows {
                     &[(1, view)],
                     "contact shadows depth",
                     layout,
-                    entries,
+                    &entries,
                 ) else {
                     return;
                 };